+++ polymorphic_tree/managers.py
@@ -1,12 +1,22 @@
 """
 The manager class for the CMS models
 """
+import django
+from django.db.models.query import QuerySet
 from mptt.managers import TreeManager
 from polymorphic import PolymorphicManager
 from polymorphic.query import PolymorphicQuerySet
 
+try:
+    # mptt 0.7 has queryset methods too
+    from mptt.querysets import TreeQuerySet
+except ImportError:
+    # provide compatibility with older mptt versions by adding a stub.
+    class TreeQuerySet(QuerySet):
+        pass
 
-class PolymorphicMPTTQuerySet(PolymorphicQuerySet):
+
+class PolymorphicMPTTQuerySet(TreeQuerySet, PolymorphicQuerySet):
     """
     Base class for querysets
     """
@@ -24,6 +34,14 @@ class PolymorphicMPTTModelManager(TreeManager, PolymorphicManager):
     #: The queryset class to use.
     queryset_class = PolymorphicMPTTQuerySet
 
+    # Re-apply the logic from django-polymorphic and django-mptt.
+    # As of django-mptt 0.7, TreeManager.get_querset() no longer calls super()
+    def get_queryset(self):
+        return self.queryset_class(self.model, using=self._db).order_by(self.tree_id_attr, self.left_attr)
+
+    # For Django 1.5
+    if django.VERSION < (1, 7):
+        get_query_set = get_queryset
 
     def toplevel(self):
         """